% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.plan.R
\name{asn.curve}
\alias{asn.curve}
\title{Compute the average sample number of a censored repetitive sampling plan.}
\usage{
asn.curve(n, k, p, asvar)
}
\arguments{
\item{n}{Sample size}

\item{k}{Vector \eqn{(k_r,k_a)} of rejection and acceptance constants.}

\item{p}{Single value or vector of nonconforming proportion}

\item{asvar}{List with the asymptotical variance-covariance matrix of MLE estimators of location and scale lifetime distribution parameters. See \code{asympt.var()} for more details.}
}
\value{
The ASN of the sampling plan.
}
\description{
This function calculates the ASN of the sampling plan coresponding to a quality level \eqn{p}
}
\examples{
q<- 0.1
asvar<-asympt.var(q,"normal")
asn.curve(11,c(1.963213, 2.214167),0.067,asvar)
}
