% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.plan.R
\name{asympt.var}
\alias{asympt.var}
\title{Compute the asymptotical variance-covariance (AVAR) matrix of MLEs of location and scale parameters}
\usage{
asympt.var(q, dist = "normal")
}
\arguments{
\item{q}{Censoring degree (\eqn{0 \le q < 1})}

\item{dist}{Location and scale distribution name ("normal"= normal, "ev" = extreme-value).
Default value is "normal".}
}
\value{
A list composed by q (censoring degree), dist (distribution) and mat (AVAR matrix of parameters estimators computed as the inverse of the Fisher information matrix).
}
\description{
This function calculates the AVAR matrix of the maximum likelihood estimators of the location and scale
parameters of normal and extreme value distributions.
}
\examples{
asympt.var(0.5)
}
