% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.params.R
\name{beta.params}
\alias{beta.params}
\title{Obtain the parameters of Generalized Beta (GB) prior according to the information about nonconforming proportion p}
\usage{
beta.params(p, l, u, know_p)
}
\arguments{
\item{p}{Vector of acceptance and rejection quality levels, \eqn{p_alpha} and \eqn{p_beta}}

\item{l}{Upper limit of \eqn{p}}

\item{know_p}{List containing the mean and variance of p or the coefficient \eqn{\epsilon} of impartiality.
In the case of impartiality, then Pr(\eqn{p \le p_alpha})=Pr(\eqn{p \ge p_beta}).}
}
\value{
A data.frame with the following variables of the parameters of GB prior
\itemize{
\item{"p_alpha": }{Acceptable quality level (AQL)}
\item{"p_beta": }{Rejectable quality level (RQL)}
\item{"a": }{Parameter a}
\item{"b": }{Parameter b}
\item{"l": }{Lower limit of \eqn{p}}
\item{"u": }{Upper limit of \eqn{p}}
\item{"mean_p": }{Mean of \eqn{p}}
\item{"var_p": }{Variance of \eqn{p}}
\item{"cdf_left": }{Pr(\eqn{p \le p_alpha}) }
\item{"cdf_right": }{Pr(\eqn{p \ge p_beta}) }
}
}
\description{
This function calculates the parameters a and b from GB prior when \eqn{l \le p \le u}. The previous knowledge can
consists in the mean and variance of p or the coefficient \eqn{\epsilon} of impartiality.
}
\examples{
p<-c(0.00654, 0.0426)
l<- p[1]/5
u<- p[2]+(p[1]-l)

# knowledge of mean and variance of p distribution
know_p<-list(mean_p=p[1],var_p=((p[2]-p[1])/4)^2)
beta.params(p,l,u, know_p)

# knowledge of epsilon-impartiality of p distribution
know_p<-list(epsilon=0.2)
beta.params(p,l,u, know_p)
}
