% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.plan.R
\name{oc.curve}
\alias{oc.curve}
\title{Compute the operating characteristic curve of a censored repetitive sampling plan.}
\usage{
oc.curve(n, k, p, asvar)
}
\arguments{
\item{n}{Sample size}

\item{k}{Vector (k_r,k_a) of rejection and acceptance constants}

\item{p}{Single value or vector of nonconforming proportion}

\item{asvar}{List with the asymptotical variance-covariance matrix of MLE estimators of location and scale lifetime distribution parameters. See \code{asympt.var()} for more details.}
}
\value{
The probability of lot acceptance of the censored repetitive plan.
}
\description{
This function calculates the OC curve that provides the probability of lot acceptance function
}
\examples{
q<- 0.1
asvar<-asympt.var(q,"normal")
oc.curve(11,c(1.963213, 2.214167),0.067,asvar)
}
