% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.plan.R
\name{rep.plan}
\alias{rep.plan}
\title{Determine the designs of conventional censored repetitive sampling plans.}
\usage{
\method{rep}{plan}(risks, p, asvar, pg_bar = TRUE)
}
\arguments{
\item{risks}{Vector of producer and consumer maximum sampling risks}

\item{p}{Vector of acceptance and rejection quality levels}

\item{asvar}{List with the asymptotical variance-covariance matrix of MLE estimators of location and scale lifetime distribution parameters. See \code{asympt.var()} for more details.}
}
\value{
A data.frame with the following variables of the censored repetitive design of the sampling plans.
\itemize{
\item{"q": }{Censoring degree}
\item{"n": }{Sample size}
\item{"ka": }{Acceptance constant}
\item{"kr": }{Rejection constant}
\item{"termcd": }{Termination code of Newton-Raphson optimization}
\item{"message": }{Final message of Newton-Raphson optimization}
\item{"p_alpha": }{Acceptable quality level (AQL)}
\item{"p_beta": }{Rejectable quality level (RQL)}
\item{"dist": }{Distribution name}
\item{"alpha": }{Maximum producer's risk}
\item{"beta": }{Maximum consumer's risk}
\item{"asn_alpha": }{ASN at AQL}
\item{"asn_beta": }{ASN at RQL }
\item{"asn_avg": }{Average of ASN at AQL and ASN at RQL}
\item{"p_asn_max": }{Quality level in which ASN maximizes}
\item{"asn_max": }{Maximum ASN over the quality level}
}
}
\description{
This function computes the designs of censored repetitive sampling plans using conventional sampling risks
given the requirements of the producer and consumer about maximum allowable risks and quality levels
}
\examples{
risks<-c(0.05,0.10)
p<-c(0.00654, 0.0426)
q<- 0.1
asvar<-asympt.var(q,"normal")
designs<-rep.plan(risks,p,asvar)

optimal.design<-designs \%>\% group_by(q,dist,p_alpha,p_beta) \%>\%
                filter( (abs(alpha-risks[1])<1e-05) & (abs(risks[2]-beta)<1e-05) & (termcd==1)) \%>\%
                slice(which.min(asn_avg)) \%>\% arrange(q,p_alpha,p_beta) \%>\% as.data.frame()

}
